Added moltiplication and division.

+++ Perl-Interpreter/Perl-Interpreter/main.c
@@ -4,12 +4,12 @@ int main(int argc, char **argv)
 {
 	if (argc != 2) return -1;
 
-	int result;
+	float result;
 	Interpreter *intrp = interpreter_create(argv[1]);
 	if (intrp == NULL) return -2;
 
 	result = interpreter_expr(intrp);
-	printf("%d\n", result);
+	printf("%f\n", result);
 
 	system("PAUSE");
 	return 0;
